fixed load transaction

+++ src/app/Components/income/income-card/income-card.component.ts
@@ -5,6 +5,7 @@ import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
 import { TransactionService } from '../../../../Core/Service/transaction.service';
 import { IncomeCategoryService } from '../../../../Core/Service/income-category.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-income-card',
@@ -46,10 +47,24 @@ export class IncomeCardComponent implements OnInit {
   }
 
   deleteTransaction(id: number) {
-    this._TransactionService.deleteTransaction(id).subscribe(() => {
-      this._TransactionService.removeTransaction(id);
-    }, (error) => {
-      console.error('Error deleting transaction:', error);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this transaction!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this._TransactionService.deleteTransaction(id).subscribe(() => {
+          this._TransactionService.removeTransaction(id);
+          Swal.fire('Deleted!', 'Your transaction has been deleted.', 'success');
+        }, (error) => {
+          console.error('Error deleting transaction:', error);
+          Swal.fire('Error!', 'There was an error deleting the transaction.', 'error');
+        });
+      }
     });
   }
